Loop unrolling now has LICM
